refactor: move some tests
